requirements_all.txt
absl-py==2.3.0
accelerate==1.8.1
aiofiles==24.1.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
calflops==0.3.2
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
custom-rasterizer==0.1
dataclasses-json==0.6.7
Deprecated==1.2.18
diffusers==0.34.0
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.13
fastapi-cli==0.0.7
faster-coco-eval==1.6.7
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
flower==2.0.1
fsspec==2025.5.1
gradio==5.34.2
gradio_client==1.10.3
greenlet==3.2.3
groovy==0.1.2
grpcio==1.73.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.1
humanfriendly==10.0
humanize==4.12.3
-e git+https://github.com/TiberiuWSJ/thesis-backend.git@fe42fb25358c5dcc82de3a726fda39233753f46f#egg=hy3dgen&subdirectory=hunyuan3d-2
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
lazy_loader==0.4
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mesh-processor==0.0.0
mpmath==1.3.0
msvc_runtime==14.42.34433
mypy_extensions==1.1.0
narwhals==1.44.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.6
omegaconf==2.3.0
onnxruntime==1.22.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.18
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
pooch==1.8.2
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
pybind11==2.13.6
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pygltflib==1.16.4
Pygments==2.19.2
PyMatting==1.1.14
pymeshlab==2023.12.post3
PyMySQL==1.1.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
rembg==2.0.66
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
ruff==0.12.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tifffile==2025.5.10
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
trimesh==4.6.12
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.34.3
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==1.17.2
xatlas==0.0.10
zipp==3.23.0
